04 auth

+++ controllers/contacts.js
@@ -3,13 +3,25 @@ const ctrlWrapper = require("../helpers/ctrlWrapper");
 const createError = require("../helpers/createError");
 
 const listContacts = async (req, res) => {
-  const result = await Contact.find({}, "-createdAT -updatedAT");
+  const { _id: owner } = req.user;
+  const { page = 1, limit = 20, favorite } = req.query;
+
+  const skip = (page - 1) * limit;
+  const searchOptions = { owner };
+  if (favorite) {
+    searchOptions.favorite = favorite;
+  }
+  const result = await Contact.find(searchOptions, "-createdAT -updatedAT", {
+    skip,
+    limit,
+  }).populate("owner", "email subscription");
   res.json(result);
 };
 
 const getContactById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
+  const result = await Contact.findById({ _id: contactId, owner });
   if (!result) {
     throw createError(404, "Not found");
   }
@@ -17,13 +29,21 @@ const getContactById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-  const result = await Contact.create(req.body);
+  const { _id: owner } = req.user;
+  const { name, phoneNumber } = req.body;
+  const existingContact = await Contact.findOne({ name, phoneNumber, owner });
+
+  if (existingContact) {
+    throw createError(409, "Contact already exists");
+  }
+  const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
 
 const removeContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndDelete(contactId);
+  const result = await Contact.findByIdAndDelete({ _id: contactId, owner });
   if (!result) {
     throw createError(404, "Not found");
   }
@@ -32,10 +52,15 @@ const removeContact = async (req, res) => {
 
 const updateContact = async (req, res) => {
   const { contactId } = req.params;
-
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
+  const { _id: owner } = req.user;
+  const result = await Contact.findByIdAndUpdate(
+    { _id: contactId, owner },
+    req.body,
+    {
+      new: true,
+      runValidators: true,
+    }
+  );
 
   if (!result) {
     throw createError(404, "Not found");
@@ -45,10 +70,11 @@ const updateContact = async (req, res) => {
 };
 
 const updateStatusContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
   const { favorite } = req.body;
   const result = await Contact.findByIdAndUpdate(
-    contactId,
+    { _id: contactId, owner },
     { favorite },
     {
       new: true,
